Move pages, fix typeError and add no covers msg
